# wfc/canvas/static/scripts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ExportArtifactsRequest": {
        "properties": {
          "file_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Types"
          },
          "run_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Ids"
          }
        },
        "title": "ExportArtifactsRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PipelineInput": {
        "properties": {
          "explicit_combos": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explicit Combos"
          },
          "keep_going": {
            "default": false,
            "title": "Keep Going",
            "type": "boolean"
          },
          "links": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PipelineLink"
            },
            "title": "Links",
            "type": "array"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/PipelineNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "param_sets": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Param Sets"
          },
          "samples": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Samples",
            "type": "array"
          },
          "variables": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variables"
          }
        },
        "required": [
          "nodes"
        ],
        "title": "PipelineInput",
        "type": "object"
      },
      "PipelineLink": {
        "properties": {
          "source": {
            "title": "Source",
            "type": "string"
          },
          "sourceHandle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcehandle"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "targetHandle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targethandle"
          }
        },
        "required": [
          "source",
          "target"
        ],
        "title": "PipelineLink",
        "type": "object"
      },
      "PipelineNode": {
        "properties": {
          "fan_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fan Mode"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "method": {
            "default": "",
            "title": "Method",
            "type": "string"
          },
          "module": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module"
          },
          "params": {
            "additionalProperties": true,
            "default": {},
            "title": "Params",
            "type": "object"
          },
          "position": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "samples": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Samples"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "method",
            "title": "Type"
          }
        },
        "required": [
          "id"
        ],
        "title": "PipelineNode",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Workflow Canvas",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/contracts/{method_full}/output_columns": {
      "get": {
        "description": "Resolve declared output columns for a method's slot.\n\nThe canvas inspector calls this for params declared with\n``column_of_input: <slot>`` to populate a dropdown of candidate column\nnames. Reuses ``wfc.contracts.resolve_columns`` against the upstream\nmethod's contract.\n\nArgs:\n    method_full: ``module.method`` (e.g. ``data_preprocessing.regionprops``).\n    slot: Output slot name to resolve columns for.\n    params: JSON-encoded dict of upstream node's current canvas params\n        (used by ``from_params`` expansion). Optional; missing/empty\n        falls back to ``{}``.\n    run_id: When the upstream is a ``run_reference`` node, pass the\n        referenced run id; the endpoint then uses Run.params from the\n        DB as the resolution context.\n\nReturns:\n    ``{strict, from_params, patterns, all}`` where ``all`` is the\n    union of resolved column names produced by ``resolve_columns``.\n    ``patterns`` are returned literally (resolution requires a CSV;\n    not done here per the no-introspection constraint).",
        "operationId": "get_output_columns_api_contracts__method_full__output_columns_get",
        "parameters": [
          {
            "in": "path",
            "name": "method_full",
            "required": true,
            "schema": {
              "title": "Method Full",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "slot",
            "required": true,
            "schema": {
              "title": "Slot",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "params",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Params"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Output Columns"
      }
    },
...
    "/api/wfc/export-artifacts": {
      "post": {
        "description": "Zip download of wfc run artifacts, optionally filtered by file type.",
        "operationId": "export_wfc_artifacts_api_wfc_export_artifacts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportArtifactsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Wfc Artifacts"
      }
    },
...
    "/api/wfc/preview-artifacts": {
      "post": {
        "description": "Preview what would be exported (all file types) without downloading.",
        "operationId": "preview_wfc_artifacts_api_wfc_preview_artifacts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportArtifactsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Wfc Artifacts"
      }
    },
...
    "/api/wfc/run/{run_id}/artifact/{artifact_path}": {
      "get": {
        "operationId": "get_wfc_artifact_api_wfc_run__run_id__artifact__artifact_path__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_path",
            "required": true,
            "schema": {
              "title": "Artifact Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Wfc Artifact"
      }
    },
    "/api/wfc/run/{run_id}/artifacts": {
      "get": {
        "operationId": "list_wfc_artifacts_api_wfc_run__run_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Wfc Artifacts"
      }
    },
...
    "/api/workflow/validate": {
      "post": {
        "description": "Validate a pipeline graph against registered methods in the DB.\n\nThe thin wrapper over the Graph unit's structural core: fetch the\ncontract map through Registration and hand the request document to\n``wfc.graph.validate_structure``, which answers ``{valid, errors,\nwarnings}``.",
        "operationId": "validate_workflow_api_workflow_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Workflow"
      }
    },
...
  }
}